feat: add partial finnish, donate section image
