+++ app/containers/DonatePage/DonatePageWrapper.js
@@ -2,6 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Flex } from 'rebass';
 import HeadTags from 'components/HeadTags';
+import { imagePathForFile } from 'utils/image';
+
+const [img] = imagePathForFile({ path: ['shutterstock', 'iwd'] });
 
 function DonatePageWrapper({ children }) {
   return (
@@ -15,6 +18,21 @@ function DonatePageWrapper({ children }) {
       flexWrap="wrap"
       m="auto"
       py={5}
+      sx={{
+        ':before': {
+          position: 'absolute',
+          top: '0',
+          left: '0',
+          width: '100%',
+          height: '100%',
+          zIndex: -1,
+          backgroundPosition: 'center',
+          content: '""',
+          backgroundImage: `linear-gradient(to bottom, #ffffffee, #ffffffcc), url(${img})`,
+          backgroundSize: 'cover',
+          filter: 'hue-rotate(310deg) saturate(1)',
+        },
+      }}
     >
       <HeadTags
         title="Donate"
