+++ app/translations/es.json
@@ -48,7 +48,7 @@
   "app.containers.HomePage.header": "The Girl Manual",
   "app.containers.HomePage.headline": "Feminismo, amistad, igualdad<br/>y otras cosas de mujer.",
   "app.containers.HomePage.hostsHeadline": "“The Girl Manual” es un podcast presentado por Fernanda Zamora y Essi Kiiski.",
-  "app.containers.HomePage.listenCta": "Escucha ahora",
+  "app.containers.HomePage.listenCta": "Escúchanos",
   "app.containers.HomePage.meetTheHosts": "Conoce a las presentadoras",
   "app.containers.LocalePicker.en": "English",
   "app.containers.LocalePicker.es": "Español",
