+++ app/translations/fi.json
@@ -1,19 +1,19 @@
 {
-  "app.components.Footer.footerMessage": "",
+  "app.components.Footer.footerMessage": "2020 Interactive Coolture - The Girl Manual",
   "app.components.Footer.patreonCta": "",
-  "app.components.Nav.header": "",
-  "app.components.NavItem.about": "",
-  "app.components.NavItem.contact": "",
-  "app.components.NavItem.donate": "",
-  "app.components.NavItem.episodes": "",
-  "app.components.NavItem.home": "",
-  "app.components.NavItem.meet": "",
+  "app.components.Nav.header": "The Girl Manual",
+  "app.components.NavItem.about": "Mikä podcasti?",
+  "app.components.NavItem.contact": "Yhteystiedot",
+  "app.components.NavItem.donate": "Lahjoita",
+  "app.components.NavItem.episodes": "Jaksot",
+  "app.components.NavItem.home": "Etusivu",
+  "app.components.NavItem.meet": "Tapaa meidät",
   "app.components.Notification.emailCheck": "",
   "app.components.Notification.emailSendError": "",
   "app.components.Notification.processingRequest": "",
   "app.containers.AboutPage.mission": "",
-  "app.containers.AboutPage.motivation": "",
-  "app.containers.AboutPage.title": "",
+  "app.containers.AboutPage.motivation": "The Girl Manual on turvallinen paikka keskustella kaikista niistä asioista, jotka liittyvät  naisena olemiseen. Feminismistä työpaikoille, naisten välisen ystävyyden ja deittailun kautta sukupuolistereotypioihin ja perheväkivaltaan. Podcast, blogi, yhteisö ja tietokeskittymä kaikesta, mikä liittyy naiseuden kokemukseen.",
+  "app.containers.AboutPage.title": "Meidän missiomme",
   "app.containers.ConfirmationPage.header": "",
   "app.containers.ContactPage.contact.email": "",
   "app.containers.ContactPage.contact.header": "",
@@ -33,22 +33,22 @@
   "app.containers.DonatePage.donate.text": "",
   "app.containers.EpisodeDetailPage.header": "",
   "app.containers.EpisodesPage.header": "",
-  "app.containers.HomePage.essi.bio": "",
-  "app.containers.HomePage.essi.location": "",
-  "app.containers.HomePage.essi.loves": "",
-  "app.containers.HomePage.essi.powers": "",
-  "app.containers.HomePage.essi.title": "",
-  "app.containers.HomePage.essi.type": "",
-  "app.containers.HomePage.fernanda.bio": "",
-  "app.containers.HomePage.fernanda.location": "",
-  "app.containers.HomePage.fernanda.loves": "",
-  "app.containers.HomePage.fernanda.powers": "",
-  "app.containers.HomePage.fernanda.title": "",
-  "app.containers.HomePage.fernanda.type": "",
-  "app.containers.HomePage.header": "",
-  "app.containers.HomePage.headline": "",
-  "app.containers.HomePage.hostsHeadline": "",
-  "app.containers.HomePage.listenCta": "",
+  "app.containers.HomePage.essi.bio": "Essi on politiikan tutkimuksen kandidaatti ja esittävän säveltaiteen maisteri. Hän työskentelee parhaillaan Brysselissä EU-päätöksenteon keskiössä ja on seurannut aiheita sosiaali- ja terveyspolitiikasta tasa-arvoon ja naisten oikeuksiin. Ennen uranvaihdostaan hän soitti freelancer-viulistina Lontoon philharmonisessa orkesterissa työskenteli viulunsoiton opettajana.",
+  "app.containers.HomePage.essi.location": "Bryssel",
+  "app.containers.HomePage.essi.loves": "Rakastaa: musiikkia + tanssia + suklaata",
+  "app.containers.HomePage.essi.powers": "Vahvuus: yleisen politiikan kandidaatti + esittävän säveltaiteen maisteri",
+  "app.containers.HomePage.essi.title": "Politiikan tutkija",
+  "app.containers.HomePage.essi.type": "Aamuvirkku",
+  "app.containers.HomePage.fernanda.bio": "Fernanda on yleisen politiikan maisteri Pariisin Science Po:sta, ja viestinnän kandidaatti Universidad Iberoamericanasta (CDMX). Hän on työskennellyt useissa projekteissa viestintäkonsulttina ja perehtynyt innovaatiopolitiikkaan, restoratiiviseen oikeuteen ja sukupuolentutkimukseen.",
+  "app.containers.HomePage.fernanda.location": "Pariisi",
+  "app.containers.HomePage.fernanda.loves": "Rakastaa: kirjallisuutta + ranskalaisia + balettia",
+  "app.containers.HomePage.fernanda.powers": "Vahvuus: yleisen politiikan maisteri + viestinnän kandidaatti",
+  "app.containers.HomePage.fernanda.title": "Kirjoittaja",
+  "app.containers.HomePage.fernanda.type": "Yökyöpeli",
+  "app.containers.HomePage.header": "The Girl Manual",
+  "app.containers.HomePage.headline": "Feminismiä, ystävyyttä, tasa-arvoa<br/>ja muita tyttöjen juttuja.",
+  "app.containers.HomePage.hostsHeadline": "“The Girl Manual” podcastia vetävät Fernanda Zamora ja Essi Kiiski",
+  "app.containers.HomePage.listenCta": "Kuuntele nyt",
   "app.containers.HomePage.meetTheHosts": "",
   "app.containers.LocalePicker.en": "",
   "app.containers.LocalePicker.es": "",
